Keep only relevant test in system parsing, move other to end-to-end
